Get rid of ProjectMeta component by moving it into Layout
